+++ mlxp/data_structures/dataframe.py
@@ -301,9 +301,10 @@ class DataFrame(list):
             if self.pandas is None:
                 self.pandas = pd.DataFrame([config._lazy() for config in self])
             return self.pandas
-    def merge(self, new_df: DataFrame)->DataFrame:
+
+    def merge(self, new_df: DataFrame) -> DataFrame:
         """Merge a target dataframe to the current dataframe.
-            
+
         The target dataframe must have the same number of rows as the current one.
 
         :params new_df: The target dataframe to merge
@@ -312,7 +313,7 @@ class DataFrame(list):
         :rtype: DataFrame
         """
 
-        for new_dict,old_dict in zip(new_df, self):
+        for new_dict, old_dict in zip(new_df, self):
             old_dict.update(new_dict)
         return DataFrame([old_dict for old_dict in self])