+++ mlxp/reader.py
@@ -96,10 +96,9 @@ class Reader(object):
         :rtype: int
         """
         return len(self.runs)
-    
+
     def __del__(self):
         self.db.close()
-
 
     def filter(
         self, query_string: str = "", result_format: str = DataFrameType.DataFrame.value,
@@ -205,12 +204,8 @@ def _get_metrics_dir(r, src_dir):
 
 
 def _get_log_dir(r, src_dir):
-    #abs_metrics_dir = r["info.logger.metrics_dir"]
-    return os.path.join(src_dir, str(r['info.logger.log_id']))
-    
-
-
-
+    # abs_metrics_dir = r["info.logger.metrics_dir"]
+    return os.path.join(src_dir, str(r["info.logger.log_id"]))
 
 
 def _get_data(path, metadata_file):