+++ CHANGELOG.md
@@ -8,5 +8,6 @@
 - Fix broken import of logs in the reader due to incorrect path.
 - Atomic logging of artifacts to avoid corrupted object when code is interupted unexpectedly.
 - Warning when using the version manager: relative paths in the config are wrt to the backup directory. To avoid this behavior use an absolute path.
+- Removing scheduler from default mlxp config.